store settings completed
